// src/teamcity-client/models/vcs-root.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TeamCity REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2018.1 (current)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
// May contain unused imports in some cases
// @ts-ignore
import type { Items } from './items';
// May contain unused imports in some cases
// @ts-ignore
import type { Project } from './project';
// May contain unused imports in some cases
// @ts-ignore
import type { Properties } from './properties';
// May contain unused imports in some cases
// @ts-ignore
import type { VcsRootInstances } from './vcs-root-instances';

/**
 * A VCS root stores repository connection settings. TeamCity uses VCS roots to create [VcsRootInstance](vcsRootInstance.md) objects, enabling build configurations and features to interact with remote repositories for source checkout, build status publishing, and more.
 * @export
 * @interface VcsRoot
 */
export interface VcsRoot {
  /**
   * The unique ID of the root. Typically consists of trunkated parent project ID and root name.
   * @type {string}
   * @memberof VcsRoot
   */
  id?: string;
  /**
   * The internal read-only ID of the root.
   * @type {string}
   * @memberof VcsRoot
   */
  internalId?: string;
  /**
   * The internal read-only UUID of the root.
   * @type {string}
   * @memberof VcsRoot
   */
  uuid?: string;
  /**
   * The public root name displayed in TeamCity UI.
   * @type {string}
   * @memberof VcsRoot
   */
  name?: string;
  /**
   * The type of a VCS provider to which this root connects.
   * @type {string}
   * @memberof VcsRoot
   */
  vcsName?: string;
  /**
   * The interval (in seconds) for polling VCS for new changes. Equals **null** if this root uses default server polling interval.
   * @type {number}
   * @memberof VcsRoot
   */
  modificationCheckInterval?: number;
  /**
   * The short (without the TeamCity server address) link to this VCS root.
   * @type {string}
   * @memberof VcsRoot
   */
  href?: string;
  /**
   *
   * @type {Project}
   * @memberof VcsRoot
   */
  project?: Project;
  /**
   *
   * @type {Properties}
   * @memberof VcsRoot
   */
  properties?: Properties;
  /**
   *
   * @type {VcsRootInstances}
   * @memberof VcsRoot
   */
  vcsRootInstances?: VcsRootInstances;
  /**
   *
   * @type {Items}
   * @memberof VcsRoot
   */
  repositoryIdStrings?: Items;
  /**
   * The part of the `href` property value that specifies the object locator. Ommitted from regular responses.
   * @type {string}
   * @memberof VcsRoot
   */
  locator?: string;
  /**
   * This property is deprecated.
   * @type {string}
   * @memberof VcsRoot
   */
  projectLocator?: string;
}